+++ HISTORY.rst
@@ -23,6 +23,8 @@ latest (unreleased)
 
 **Bugfixes**
 
+* BIZ-1067: Define a default paperformat when company is created
+
 **Build**
 
 **Documentation**